+++ test_model01/generator.cpp
@@ -1,18 +1,18 @@
 #include "StdAfx.h"
 #include "generator.h"
 
-double summarize(vector<double> values)
+float summarize(vector<float> values)
 {
-	double sum = 0;
+	float sum = 0;
 	for (int i = 0; i < values.size(); i++)
 		sum += values[i];
 	return sum;
 }
 
-vector<double> allocate(vector<double> values)
+vector<float> allocate(vector<float> values)
 {
-	vector<double> allocation;
-	double sum = summarize(values);
+	vector<float> allocation;
+	float sum = summarize(values);
 	for (int i = 0; i < values.size(); i++)
 	{
 		allocation.push_back(values[i]/sum);
@@ -20,9 +20,9 @@ vector<double> allocate(vector<double> values)
 	return allocation;
 }
 
-vector<double> invert(vector<double> values)
+vector<float> invert(vector<float> values)
 {
-	vector<double> invert_vector;
+	vector<float> invert_vector;
 	for (int i = 0; i < values.size(); i++)
 	{
 		if (values[i])
@@ -33,15 +33,15 @@ vector<double> invert(vector<double> values)
 	return invert_vector;
 }
 
-int get_random(vector<double> probabilities)
+int get_random(vector<float> probabilities)
 {
-	vector<double> allocation;
+	vector<float> allocation;
 	allocation.push_back(0);
 	for (int i = 0; i < probabilities.size(); i++)
 	{
 		allocation.push_back(allocation[i] + probabilities[i]);
 	}
-	double rand_number = rand()/(double)RAND_MAX;
+	float rand_number = rand()/(float)RAND_MAX;
 	if (rand_number == 1)
 		return (allocation.size() - 1);
 	for (int i = 1; i < allocation.size(); i++)
