+++ test_model01/generator.h
@@ -5,57 +5,57 @@
 
 using namespace std;
 
-double summarize(vector<double> values);
+float summarize(vector<float> values);
 
-vector<double> allocate(vector<double> values);	
+vector<float> allocate(vector<float> values);	
 
-template <class T> map<T, double> allocate(map <T,double> values);	
+template <class T> map<T, float> allocate(map <T,float> values);	
 
-vector<double> invert(vector<double> values);
+vector<float> invert(vector<float> values);
 
-template <class T> map<T,double> invert(map <T,double> values);
+template <class T> map<T,float> invert(map <T,float> values);
 
-int get_random(vector<double> probabilities);
+int get_random(vector<float> probabilities);
 
-template <class T> T get_random(map<T,double> probabilities);
+template <class T> T get_random(map<T,float> probabilities);
 
-template <class T> double summarize(map<T, double> values)
+template <class T> float summarize(map<T, float> values)
 {
-	double sum = 0;
-	for (map<T, double>::iterator i = values.begin(); i != values.end(); ++i)
+	float sum = 0;
+	for (map<T, float>::iterator i = values.begin(); i != values.end(); ++i)
 		sum += i->second;
 	return sum;
 }
 
-template <class T> map<T, double> allocate(map<T, double> values)
+template <class T> map<T, float> allocate(map<T, float> values)
 {
-	map<T, double> allocation;
-	double sum = summarize<T>(values);
-	for (map<T, double>::iterator i = values.begin(); i != values.end(); ++i)
+	map<T, float> allocation;
+	float sum = summarize<T>(values);
+	for (map<T, float>::iterator i = values.begin(); i != values.end(); ++i)
 	{
 		allocation[i->first] = (i->second/sum);
 	}
 	return allocation;
 }
 
-template <class T> map<T, double> invert(map<T, double> values)
+template <class T> map<T, float> invert(map<T, float> values)
 {
-	map<T, double> invert_vector;
-	for (map<T, double>::iterator i = values.begin(); i != values.end(); ++i)
+	map<T, float> invert_vector;
+	for (map<T, float>::iterator i = values.begin(); i != values.end(); ++i)
 	{
 		invert_vector[i->first] = (1.0/i->second);
 	}
 	return invert_vector;
 }
 
-template <class T> T get_random(map<T, double> probabilities)
+template <class T> T get_random(map<T, float> probabilities)
 {
-	double rand_number = rand()/(double)RAND_MAX;
+	float rand_number = rand()/(float)RAND_MAX;
 	if (rand_number == 1)
 		return (probabilities.rbegin()->first);
-	map<T, double>::iterator index = probabilities.begin();
-	double first = 0;
-	double last = index->second;
+	map<T, float>::iterator index = probabilities.begin();
+	float first = 0;
+	float last = index->second;
 	while (index != probabilities.end()) 
 	{
 		if (first <= rand_number && rand_number < last)
